Add 104-fibonacci.c

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -8,13 +8,13 @@
 int main(void)
 {
 	int n = 98, i;
-	unsigned long fib1 = 1, fib2 = 2, fib;
+	int fib1 = 1, fib2 = 2, fib;
 
-	printf("%lu, %lu", fib1, fib2);
+	printf("%d, %d", fib1, fib2);
 	for (i = 3; i <= n; i++)
 	{
 		fib = fib1 + fib2;
-		printf(", %lu", fib);
+		printf(", %d", fib);
 		fib1 = fib2;
 		fib2 = fib;
 	}
